Refactor of ImportContactsFeedbackTest
