+++ qabelbox/src/androidTest/java/de/qabel/qabelbox/ui/files/FilesFragmentUITest.java
@@ -169,7 +169,6 @@ public class FilesFragmentUITest {
     }
 
     @Test
-    @Ignore("Drop messages broken")
     public void shareFileTest() {
         Spoon.screenshot(mActivity, "startup");
         onView(withText(exampleFiles.get(0).getName())).perform(longClick());
