+++ qabelbox/src/main/java/de/qabel/qabelbox/activities/MainActivity.java
@@ -117,9 +117,10 @@ public class MainActivity extends CrashReportingActivity
     private static final int REQUEST_CODE_OPEN = 21;
     private static final int REQUEST_CODE_DELETE_FILE = 22;
 
-    // Intent extra to specifiy if the files fragment should be started
+    // Intent extra to specify if the files fragment should be started
     // Defaults to true and is used in tests to shortcut the activity creation
     public static final String START_FILES_FRAGMENT = "START_FILES_FRAGMENT";
+    public static final String START_CONTACTS_FRAGMENT = "START_CONTACTS_FRAGMENT";
 
     private DrawerLayout drawer;
     public BoxVolume boxVolume;
@@ -419,9 +420,9 @@ public class MainActivity extends CrashReportingActivity
         // Checks if a fragment should be launched
 
         boolean start_files_fragment = intent.getBooleanExtra(START_FILES_FRAGMENT, true);
-        if (type != null && intent != null && intent.getAction() != null) {
+        boolean start_contacts_fragment = intent.getBooleanExtra(START_CONTACTS_FRAGMENT, false);
+        if (type != null && intent.getAction() != null) {
             String scheme = intent.getScheme();
-
 
             switch (intent.getAction()) {
                 case Intent.ACTION_VIEW:
@@ -449,13 +450,17 @@ public class MainActivity extends CrashReportingActivity
                     }
                     break;
                 default:
-                    if (start_files_fragment) {
+                    if (start_contacts_fragment) {
+                        selectContactsFragment();
+                    } else if (start_files_fragment) {
                         initAndSelectFilesFragment();
                     }
                     break;
             }
         } else {
-            if (start_files_fragment) {
+            if (start_contacts_fragment) {
+                selectContactsFragment();
+            } else if (start_files_fragment) {
                 initAndSelectFilesFragment();
             }
         }
